chatbot/be-chatbot-educare-master/requirements.txt
django 
djangorestframework 
django-cors-headers 
nltk 
keras 
deep-translator 
langdetect 
underthesea
tensorflow